role-permission Sekolah
menerapkan role-permission pada data Sekolah

+++ app/Http/Controllers/SekolahController.php
@@ -6,18 +6,34 @@ use App\Http\Requests\StoreSekolahRequest;
 use App\Http\Requests\UpdateSekolahRequest;
 use App\Models\Sekolah;
 use Illuminate\Http\Request;
+use Yajra\DataTables\Facades\DataTables;
 
 class SekolahController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $sekolahs = Sekolah::query()->latest()->get();
-
-        return view('sekolah', compact('sekolahs'));
+        if ($request->ajax()) {
+            $sekolah = Sekolah::query()->latest();
+            
+            return DataTables::of($sekolah)
+                ->addIndexColumn()
+                ->addColumn('action', function ($sekolah) {
+                    return '
+                        <a href="'.route('sekolah.edit', $sekolah->id).'" class="text-blue-600">Edit</a>
+                        <form action="'.route('sekolah.destroy', $sekolah->id).'" method="POST" style="display:inline-block;" onsubmit="return confirm(\'Yakin akan menghapus data?\')">
+                            '.csrf_field().method_field('DELETE').'
+                            <button type="submit" class="text-red-600">Delete</button>
+                        </form>';
+                })
+                ->rawColumns(['action'])
+                ->make(true);
+        }
+        return view('sekolah');
     }
+
 
     /**
      * Show the form for creating a new resource.
@@ -68,12 +84,16 @@ class SekolahController extends Controller
      */
     public function destroy(Sekolah $sekolah)
     {
+        if (!auth()->user()->can('hapus-jurusan')) {
+            return redirect()->route('sekolah.index')->with('error', 'Anda tidak memiliki izin untuk menghapus data sekolah.');
+        }
+
         if ($sekolah->jurusanSekolahs()->exists()) {
             return redirect()->back()->with('error', 'Tidak dapat menghapus data karena terdapat relasi dengan data lain.');
         }
 
         $sekolah->delete();
-    
         return redirect()->route('sekolah.index')->with('success', 'Data berhasil dihapus.');
     }
+
 }
